refactor(pr): Add utility to lookup PR from number or revision
